fixed issues

+++ Node.h
@@ -2,7 +2,6 @@
 #include <iostream>
 #include <string>
 #include "doubly_linked_list.h"
-#include"Edge.h"
 #include"graph.h"
 using namespace std;
 class NODE;
@@ -15,7 +14,7 @@ public:
 	NODE* prev;
 	int dist_from_the_first_node;
 
-	NODE(int id ) : id(id) , prev(NULL) , dist_from_the_first_node(INT_MAX) 
+	NODE(int id ) : id(id) , prev(NULL) , dist_from_the_first_node(INT16_MAX) 
 	{
 		nodes.insert(this);
 	}
